[mlir][vector] Allow unroll of contraction in arbitrary order
Adds a functor that assists with index calculations while unrolling a
vector operation. Vector unroll/distribute transformations can then
use this functor as a drop-in replacement to enable unrolling in different
orders. For example, the `vector.contract` may be decomposed into a
smaller set of contractions. There is a choice of how to unroll
the decomposition based on the traversal order of (dim0, dim1, dim2).
The choice of traversal order is then given by a callback which can be
specified by the user of the transform. For now, only the `vector.contract`,
`vector.transfer_read` and `vector.transfer_write` utilize this callback.